% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-import-html-into-r.R
\name{read_html_table}
\alias{read_html_table}
\title{Read HTML table and render it nicely with Quarto}
\usage{
read_html_table(ctry, table = 1L)
}
\arguments{
\item{ctry}{country code}

\item{table}{table number (default: 1)}
}
\value{
HTML table
}
\description{
Read HTML table and render it nicely with Quarto
}
