% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-import-html-into-r.R
\name{read_html_text}
\alias{read_html_text}
\title{Read HTML policy development text and render it nicely with Quarto}
\usage{
read_html_text(ctry)
}
\arguments{
\item{ctry}{country code}
}
\value{
Text for policy developments
}
\description{
Read HTML policy development text and render it nicely with Quarto
}
